+++ docs/release-notes.md
@@ -3,6 +3,7 @@
 - General: UI can be hidden when launching a custom browser (useful for terminal commands).
 - Toast: notification shows browser icon instead of BT logo.
 - Toast: visibility interval is configurable in settings.
+- Toast: border width is configurable in settings.
 - Picker: added option to customise window border width (previously there was no border, now will default to 1).
 - Picker: mouse cursor changes to "hand" on browser icon hover.
 - Picker: option to draw native Windows chrome (off by default).